+++ sdk/messaging/azeventgrid/main_test.go
@@ -15,14 +15,30 @@ import (
 	"github.com/joho/godotenv"
 )
 
-func TestMain(m *testing.M) {
-	if err := recording.ResetProxy(nil); err != nil {
-		panic(err)
-	}
+const recordingDirectory = "sdk/messaging/azeventgrid/testdata"
 
+func TestMain(m *testing.M) {
+	code := run(m)
+	os.Exit(code)
+}
+
+func run(m *testing.M) int {
+	if recording.GetRecordMode() == recording.PlaybackMode || recording.GetRecordMode() == recording.RecordingMode {
+		proxy, err := recording.StartTestProxy(recordingDirectory, nil)
+		if err != nil {
+			panic(err)
+		}
+
+		defer func() {
+			err := recording.StopTestProxy(proxy)
+			if err != nil {
+				panic(err)
+			}
+		}()
+	}
 	if err := godotenv.Load(); err != nil {
 		log.Printf("Failed to load .env file, no integration tests will run: %s", err)
 	}
 
-	os.Exit(m.Run())
+	return m.Run()
 }