+++ sdk/messaging/azeventhubs/consumer_client_test.go
@@ -882,9 +882,7 @@ func mustSendEventsToAllPartitions(t *testing.T, events []*azeventhubs.EventData
 		}(partitionID)
 	}
 
-	t.Logf("Sending all events to %d partitions", len(hubProps.PartitionIDs))
 	wg.Wait()
-	t.Logf("Done sending events")
 	close(partitionsCh)
 
 	var partitions []azeventhubs.PartitionProperties
@@ -920,8 +918,6 @@ func getSortedBodies(events []*azeventhubs.ReceivedEventData) []string {
 	return bodies
 }
 
-const destPartitionIDKey = "DestPartitionID"
-
 func sendEventToPartition(t *testing.T, producer *azeventhubs.ProducerClient, partitionID string, events []*azeventhubs.EventData) azeventhubs.PartitionProperties {
 	partProps, err := producer.GetPartitionProperties(context.Background(), partitionID, nil)
 	require.NoError(t, err)
@@ -936,7 +932,7 @@ func sendEventToPartition(t *testing.T, producer *azeventhubs.ProducerClient, pa
 		eventToSend := *event
 
 		props := map[string]any{
-			destPartitionIDKey: partitionID,
+			"DestPartitionID": partitionID,
 		}
 
 		for k, v := range event.Properties {
@@ -945,7 +941,7 @@ func sendEventToPartition(t *testing.T, producer *azeventhubs.ProducerClient, pa
 
 		eventToSend.Properties = props
 
-		err = batch.AddEventData(&eventToSend, nil)
+		err = batch.AddEventData(event, nil)
 		require.NoError(t, err)
 	}