+++ sdk/security/keyvault/azsecrets/utils_test.go
@@ -27,6 +27,7 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+const recordingDirectory = "sdk/security/keyvault/azsecrets/testdata"
 const fakeVaultURL = "https://fakevault.local"
 
 var (
@@ -40,6 +41,25 @@ var (
 )
 
 func TestMain(m *testing.M) {
+	code := run(m)
+	os.Exit(code)
+}
+
+func run(m *testing.M) int {
+	if recording.GetRecordMode() == recording.PlaybackMode || recording.GetRecordMode() == recording.RecordingMode {
+		proxy, err := recording.StartTestProxy(recordingDirectory, nil)
+		if err != nil {
+			panic(err)
+		}
+
+		defer func() {
+			err := recording.StopTestProxy(proxy)
+			if err != nil {
+				panic(err)
+			}
+		}()
+	}
+
 	vaultURL = fakeVaultURL
 	if recording.GetRecordMode() != recording.PlaybackMode {
 		if u, ok := os.LookupEnv("AZURE_KEYVAULT_URL"); ok && u != "" {
@@ -48,16 +68,13 @@ func TestMain(m *testing.M) {
 			panic("no value for AZURE_KEYVAULT_URL")
 		}
 	}
-	proxy, err := recording.StartTestProxy("sdk/security/keyvault/azsecrets/testdata", nil)
-	if err != nil {
-		panic(err)
-	}
 	if recording.GetRecordMode() == recording.PlaybackMode {
 		credential = &FakeCredential{}
 	} else {
 		tenantID := lookupEnvVar("AZSECRETS_TENANT_ID")
 		clientID := lookupEnvVar("AZSECRETS_CLIENT_ID")
 		secret := lookupEnvVar("AZSECRETS_CLIENT_SECRET")
+		var err error
 		credential, err = azidentity.NewClientSecretCredential(tenantID, clientID, secret, nil)
 		if err != nil {
 			panic(err)
@@ -76,12 +93,6 @@ func TestMain(m *testing.M) {
 		if err != nil {
 			panic(err)
 		}
-		defer func() {
-			err := recording.ResetProxy(nil)
-			if err != nil {
-				panic(err)
-			}
-		}()
 	}
 	code := m.Run()
 	if recording.GetRecordMode() != recording.PlaybackMode {
@@ -110,15 +121,11 @@ func TestMain(m *testing.M) {
 		}
 	}
 
-	err = recording.StopTestProxy(proxy)
-	if err != nil {
-		panic(err)
-	}
-	os.Exit(code)
+	return code
 }
 
 func startTest(t *testing.T) *azsecrets.Client {
-	err := recording.Start(t, "sdk/security/keyvault/azsecrets/testdata", nil)
+	err := recording.Start(t, recordingDirectory, nil)
 	require.NoError(t, err)
 	t.Cleanup(func() {
 		err := recording.Stop(t, nil)