+++ sdk/storage/azblob/service/client_test.go
@@ -46,6 +46,14 @@ func Test(t *testing.T) {
 	}
 }
 
+func (s *ServiceRecordedTestsSuite) SetupSuite() {
+	s.proxy = testcommon.SetupSuite(&s.Suite)
+}
+
+func (s *ServiceRecordedTestsSuite) TearDownSuite() {
+	testcommon.TearDownSuite(&s.Suite, s.proxy)
+}
+
 func (s *ServiceRecordedTestsSuite) BeforeTest(suite string, test string) {
 	testcommon.BeforeTest(s.T(), suite, test)
 }
@@ -64,6 +72,7 @@ func (s *ServiceUnrecordedTestsSuite) AfterTest(suite string, test string) {
 
 type ServiceRecordedTestsSuite struct {
 	suite.Suite
+	proxy *recording.TestProxyInstance
 }
 
 type ServiceUnrecordedTestsSuite struct {