+++ sdk/tracing/azotel/otel_test.go
@@ -28,8 +28,10 @@ func TestNewTracingProvider(t *testing.T) {
 
 	otelTP := tracesdk.NewTracerProvider(tracesdk.WithBatcher(exporter))
 
-	client, err := azcore.NewClient("azotel.TestClient", internal.Version, azruntime.PipelineOptions{
-		TracingNamespace: "TestNewTracingProvider",
+	client, err := azcore.NewClient("azotel", internal.Version, azruntime.PipelineOptions{
+		Tracing: azruntime.TracingOptions{
+			Namespace: "TestNewTracingProvider",
+		},
 	}, &azcore.ClientOptions{
 		TracingProvider: NewTracingProvider(otelTP, nil),
 	})